+++ test/bigtest/network/scenarios/fetch-mapping-profiles-success.js
@@ -1,0 +1,42 @@
+import { searchEntityByQuery } from '../../helpers/searchEntityByQuery';
+
+export default server => {
+  server.createList('mapping-profile', 3);
+
+  server.get('/data-import-profiles/mappingProfiles', (schema, request) => {
+    const { query = '' } = request.queryParams;
+    const mappingProfiles = schema.mappingProfiles.all();
+
+    const searchPattern = /name="([\w\s]+)/;
+
+    return searchEntityByQuery({
+      query,
+      entity: mappingProfiles,
+      searchPattern,
+      fieldsToMatch: [
+        'name',
+        'mapped',
+        'tags.tagList',
+      ],
+    });
+  });
+  server.get('/data-import-profiles/mappingProfiles/:id');
+  server.post('/data-import-profiles/mappingProfiles', (_, request) => {
+    const params = JSON.parse(request.requestBody);
+    const record = server.create('mapping-profile', params);
+
+    return record.attrs;
+  });
+  server.put('/data-import-profiles/mappingProfiles/:id', (schema, request) => {
+    const {
+      params: { id },
+      requestBody,
+    } = request;
+    const mappingProfileModel = schema.mappingProfiles.find(id);
+    const updatedMappingProfile = JSON.parse(requestBody);
+
+    mappingProfileModel.update({ ...updatedMappingProfile });
+
+    return mappingProfileModel.attrs;
+  });
+};